# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "remex-server",
    "description": "",
    "contact": {
      "name": "Zion Koyl",
      "email": "hiszd1@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs/addclients": {
      "post": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "add_clients_to_jobs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobClientAction"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Clients added to jobs successfully"
          }
        }
      }
    },
    "/jobs/new": {
      "post": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "create_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/removeclients": {
      "post": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "remove_clients_from_jobs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobClientAction"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Clients removed from jobs successfully"
          }
        }
      }
    },
    "/jobs/update": {
      "post": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "update_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobSRV"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSRV"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "get_job_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobWithClients"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "delete_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job deleted successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/jobs/{id}/groups": {
      "get": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "get_job_groups",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groups for job found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobWithGroups"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/jobs/{job_id}/client-statuses": {
      "get": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "get_job_client_statuses",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client statuses for job found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientJobStatusResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/jobs/{job_id}/groups": {
      "post": {
        "tags": [
          "handlers::jobs"
        ],
        "operationId": "update_job_groups",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobGroupsForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job groups updated successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClientJobStatusResponse": {
        "type": "object",
        "required": [
          "client_id",
          "client_name",
          "status",
          "execution_count"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_name": {
            "type": "string"
          },
          "execution_count": {
            "type": "integer",
            "format": "int64"
          },
          "latest_execution_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "latest_execution_timestamp": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ClientSRV": {
        "type": "object",
        "required": [
          "id",
          "secret",
          "client_name",
          "hardware_hash",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "client_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "hardware_hash": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateJobForm": {
        "type": "object",
        "required": [
          "job_name",
          "job_type",
          "job_status",
          "job_shell",
          "job_command"
        ],
        "properties": {
          "frequency": {
            "type": [
              "string",
              "null"
            ]
          },
          "group_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "job_command": {
            "type": "string"
          },
          "job_name": {
            "type": "string"
          },
          "job_shell": {
            "type": "string"
          },
          "job_status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "job_type": {
            "type": "string"
          },
          "scheduled_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Job": {
        "type": "object",
        "required": [
          "id",
          "job_name",
          "job_type",
          "job_status",
          "job_shell",
          "job_command",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "job_command": {
            "type": "string"
          },
          "job_name": {
            "type": "string"
          },
          "job_shell": {
            "type": "string"
          },
          "job_status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JobClientAction": {
        "type": "object",
        "required": [
          "job_id",
          "client_id"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          }
        }
      },
...
      "JobSRV": {
        "type": "object",
        "required": [
          "id",
          "job_name",
          "job_type",
          "job_status",
          "job_shell",
          "job_command",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "job_command": {
            "type": "string"
          },
          "job_name": {
            "type": "string"
          },
          "job_shell": {
            "type": "string"
          },
          "job_status": {
            "type": "string"
          },
          "job_type": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JobStatus": {
        "oneOf": [
          {
            "type": "string",
            "description": "The job is waiting to be picked up for execution.",
            "enum": [
              "pending"
            ]
          },
          {
            "type": "string",
            "description": "The job is currently being executed by a client.",
            "enum": [
              "running"
            ]
          },
          {
            "type": "string",
            "description": "The job has finished successfully.",
            "enum": [
              "completed"
            ]
          },
          {
            "type": "object",
            "description": "The job has failed during execution, includes an error message.",
            "required": [
              "failed"
            ],
            "properties": {
              "failed": {
                "type": "string",
                "description": "The job has failed during execution, includes an error message."
              }
            }
          },
          {
            "type": "string",
            "description": "The job was cancelled by a user or system administrator.",
            "enum": [
              "cancelled"
            ]
          },
          {
            "type": "string",
            "description": "The job execution exceeded the allocated time limit.",
            "enum": [
              "timed_out"
            ]
          },
          {
            "type": "string",
            "description": "The job is disabled and will not be picked up for execution.",
            "enum": [
              "disabled"
            ]
          }
        ]
      },
      "JobType": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "instant"
            ]
          },
          {
            "type": "object",
            "required": [
              "scheduled"
            ],
            "properties": {
              "scheduled": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "recurring"
            ],
            "properties": {
              "recurring": {
                "type": "array",
                "items": {
                  "type": "object"
                },
                "maxItems": 2,
                "minItems": 2
              }
            }
          }
        ]
      },
      "JobWithClients": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JobSRV"
          },
          {
            "type": "object",
            "required": [
              "clients"
            ],
            "properties": {
              "clients": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ClientSRV"
                }
              }
            }
          }
        ]
      },
      "JobWithGroups": {
        "type": "object",
        "required": [
          "id",
          "group_name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "group_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateJobGroupsForm": {
        "type": "object",
        "required": [
          "group_ids"
        ],
        "properties": {
          "group_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateJobSRV": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "job_command": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_shell": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_status": {
            "type": [
              "string",
              "null"
            ]
          },
          "job_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      }
    }
  }
}